Common function for finding test samples
