+++ tests/integration.rs
@@ -1,3 +1,4 @@
+mod common;
 mod formatting;
 mod parsing;
 mod templating;
